dev-requirements.txt
-r requirements.txt
# nose==1.3.7
ckanapi==4.6
beautifulsoup4==4.5.1
requests-mock==1.9.3
httpretty==1.1.3
pytest-ckan
pytest-cov
